requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
altair==4.2.0
appnope==0.1.3
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
blinker==1.5
cachetools==5.2.0
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.5
cycler==0.11.0
dateparser==1.1.1
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.0.0
fonttools==4.37.3
frozenlist==1.3.1
gitdb==4.0.10
GitPython==3.1.29
idna==3.4
importlib-metadata==5.1.0
ipykernel==6.15.3
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
multidict==6.0.2
multitasking==0.0.11
nest-asyncio==1.5.5
numpy==1.23.3
packaging==21.3
pandas==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-binance==1.0.16
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==24.0.1
regex==2022.3.2
requests==2.28.1
rich==12.6.0
scikit-learn==1.1.2
scipy==1.9.1
semver==2.13.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
stack-data==0.5.0
streamlit==1.15.2
ta==0.10.2
TA-Lib==0.4.25
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.4.0
typing_extensions==4.4.0
tzdata==2022.4
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
validators==0.20.0
wcwidth==0.2.5
websockets==10.3
xgboost==1.6.2
yarl==1.8.1
yfinance==0.1.74
zipp==3.11.0
